Changes to css, images, about page and home page, and footer
CSS: added css for about page, scrolling, and home page(there was a mix
up with home_left and home_right so I put them in the right spots on
the home page and changed it’s corresponding names on the css)

Images: added images used for footer media icons and added image used
for the background of the about page

About page: did the layout for the about page and added text for about
the company. Still need to put info about the founder. Added scroll
down and up arrows

Home page: added scroll down and up arrows. Trying to figure out how to
add the calendar to the home page

Footer: Added the media icons to the footer